+++ crates/siril-sys/src/lib.rs
@@ -1,3 +1,4 @@
+mod conversion;
 mod output;
 mod siril;
 mod types;
@@ -6,6 +7,7 @@ pub mod commands;
 pub mod message;
 pub mod siril_ext;
 
+pub use conversion::{ConversionEntry, ConversionFile};
 pub use output::{OutputLine, OutputSink, OutputStream};
 pub use siril::Builder;
 pub use siril::Siril;